CV_WEEK01_2_AM/requirements.txt
Pillow==8.2.0
PySimpleGUI==4.40.0
